+++ apps/dashboard/package.json
@@ -39,9 +39,6 @@
     "@yukinu/tsconfig": "workspace:*",
     "tailwindcss": "catalog:tailwind",
     "typescript": "catalog:dev",
-    "vite": "npm:rolldown-vite@^7.3.1"
-  },
-  "overrides": {
-    "vite": "npm:rolldown-vite@^7.3.0"
+    "vite": "^8.0.0"
   }
 }
