+++ packages/auth/package.json
@@ -6,8 +6,9 @@
   "types": "./dist/index.d.mts",
   "exports": {
     ".": {
-      "react-server": "./dist/index.rsc.mjs",
-      "default": "./dist/index.mjs"
+      "types": "./dist/index.d.mts",
+      "default": "./dist/index.mjs",
+      "react-server": "./dist/index.rsc.mjs"
     },
     "./react": {
       "types": "./dist/react.d.mts",
