+++ packages/lib/package.json
@@ -3,15 +3,6 @@
   "version": "0.7.2",
   "private": true,
   "type": "module",
-  "exports": {
-    "./constants": "./dist/constants.mjs",
-    "./create-env": "./dist/create-env.mjs",
-    "./create-id": "./dist/create-id.mjs",
-    "./create-query-client": "./dist/create-query-client.mjs",
-    "./rate-limit": "./dist/rate-limit.mjs",
-    "./slugify": "./dist/slugify.mjs",
-    "./package.json": "./package.json"
-  },
   "scripts": {
     "build": "tsdown",
     "clean": "git clean -xdf .cache .turbo dist node_modules",
@@ -28,5 +19,40 @@
     "@yukinu/tsconfig": "workspace:*",
     "tsdown": "catalog:dev",
     "typescript": "catalog:dev"
+  },
+  "exports": {
+    "./constants": {
+      "types": "./dist/constants.d.mts",
+      "default": "./dist/constants.mjs"
+    },
+    "./create-env": {
+      "types": "./dist/create-env.d.mts",
+      "default": "./dist/create-env.mjs"
+    },
+    "./create-id": {
+      "types": "./dist/create-id.d.mts",
+      "default": "./dist/create-id.mjs"
+    },
+    "./create-query-client": {
+      "types": "./dist/create-query-client.d.mts",
+      "default": "./dist/create-query-client.mjs"
+    },
+    "./rate-limit": {
+      "types": "./dist/rate-limit.d.mts",
+      "default": "./dist/rate-limit.mjs"
+    },
+    "./slugify": {
+      "types": "./dist/slugify.d.mts",
+      "default": "./dist/slugify.mjs"
+    },
+    "./standard-schema": {
+      "types": "./dist/standard-schema.d.mts",
+      "default": "./dist/standard-schema.mjs"
+    },
+    "./standard-schema.d.ts": {
+      "types": "./dist/standard-schema.d.ts.d.mts",
+      "default": "./dist/standard-schema.d.ts.mjs"
+    },
+    "./package.json": "./package.json"
   }
 }
