+++ packages/lib/tsconfig.json
@@ -1,6 +1,10 @@
 {
   "extends": "@yukinu/tsconfig/internal-package.json",
-  "compilerOptions": {},
+  "compilerOptions": {
+    "paths": {
+      "@/*": ["./src/*"]
+    }
+  },
   "include": ["src"],
   "exclude": ["node_modules"]
 }
