+++ packages/ui/package.json
@@ -3,9 +3,11 @@
   "version": "0.8.4",
   "private": true,
   "type": "module",
-  "types": "./dist/index.d.mts",
   "exports": {
-    ".": "./dist/index.mjs",
+    ".": {
+      "types": "./dist/index.d.mts",
+      "default": "./dist/index.mjs"
+    },
     "./*": {
       "types": "./dist/components/*.d.mts",
       "default": "./dist/components/*.mjs"
@@ -25,6 +27,7 @@
   },
   "dependencies": {
     "@base-ui/react": "^1.3.0",
+    "@yukinu/lib": "workspace:*",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.577.0",
