+++ packages/ui/src/hooks/use-form.tsx
@@ -1,4 +1,5 @@
-// oxlint-disable max-statements
+import type { StandardSchemaV1 } from '@yukinu/lib/standard-schema'
+
 import * as React from 'react'
 
 interface FormError {
@@ -257,82 +258,4 @@ export function useForm<
     }),
     [formId, Field, handleSubmit, isPending, reset],
   )
-}
-
-/** The Standard Schema interface. */
-interface StandardSchemaV1<Input = unknown, Output = Input> {
-  /** The Standard Schema properties. */
-  readonly '~standard': StandardSchemaV1.Props<Input, Output>
-}
-
-// oxlint-disable-next-line typescript/no-namespace
-declare namespace StandardSchemaV1 {
-  /** The Standard Schema properties interface. */
-  export interface Props<Input = unknown, Output = Input> {
-    /** The version number of the standard. */
-    readonly version: 1
-    /** The vendor name of the schema library. */
-    readonly vendor: string
-    /** Validates unknown input values. */
-    readonly validate: (
-      value: unknown,
-      options?: StandardSchemaV1.Options | undefined,
-    ) => Result<Output> | Promise<Result<Output>>
-    /** Inferred types associated with the schema. */
-    readonly types?: Types<Input, Output> | undefined
-  }
-
-  /** The result interface of the validate function. */
-  export type Result<Output> = SuccessResult<Output> | FailureResult
-
-  /** The result interface if validation succeeds. */
-  export interface SuccessResult<Output> {
-    /** The typed output value. */
-    readonly value: Output
-    /** A falsy value for `issues` indicates success. */
-    readonly issues?: undefined
-  }
-
-  export interface Options {
-    /** Explicit support for additional vendor-specific parameters, if needed. */
-    readonly libraryOptions?: Record<string, unknown> | undefined
-  }
-
-  /** The result interface if validation fails. */
-  export interface FailureResult {
-    /** The issues of failed validation. */
-    readonly issues: readonly Issue[]
-  }
-
-  /** The issue interface of the failure output. */
-  export interface Issue {
-    /** The error message of the issue. */
-    readonly message: string
-    /** The path of the issue, if any. */
-    readonly path?: readonly (PropertyKey | PathSegment)[] | undefined
-  }
-
-  /** The path segment interface of the issue. */
-  export interface PathSegment {
-    /** The key representing a path segment. */
-    readonly key: PropertyKey
-  }
-
-  /** The Standard Schema types interface. */
-  export interface Types<Input = unknown, Output = Input> {
-    /** The input type of the schema. */
-    readonly input: Input
-    /** The output type of the schema. */
-    readonly output: Output
-  }
-
-  /** Infers the input type of a Standard Schema. */
-  export type InferInput<Schema extends StandardSchemaV1> = NonNullable<
-    Schema['~standard']['types']
-  >['input']
-
-  /** Infers the output type of a Standard Schema. */
-  export type InferOutput<Schema extends StandardSchemaV1> = NonNullable<
-    Schema['~standard']['types']
-  >['output']
 }
